requirements.txt
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
